test is normally not accessible from demo.
