+++ demo/oauth2_add_on_dpop.py
@@ -1,6 +1,9 @@
+#!/usr/bin/env python3
 import json
-import os
 
+from common import BASEDIR
+from common import KEYDEFS
+from common import full_path
 from flow import Flow
 from idpyoidc.claims import get_signing_algs
 from idpyoidc.client.oauth2 import Client
@@ -10,18 +13,6 @@ from idpyoidc.server.user_info import UserInfo
 from oauth2_client_conf import CLIENT_CONFIG
 from oauth2_client_conf import CLIENT_ID
 from oauth2_server_conf import SERVER_CONF
-
-KEYDEFS = [
-    {"type": "RSA", "key": "", "use": ["sig"]},
-    {"type": "EC", "crv": "P-256", "use": ["sig"]},
-]
-
-BASEDIR = os.path.abspath(os.path.dirname(__file__))
-
-
-def full_path(local_file):
-    return os.path.join(BASEDIR, local_file)
-
 
 # ================ Server side ===================================
