+++ demo/oauth2_add_on_par.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env python3
-import os
-import uuid
 
+from common import BASEDIR
+from common import KEYDEFS
 from flow import Flow
 from idpyoidc.client.oauth2 import Client
 from idpyoidc.server import Server
@@ -9,18 +9,6 @@ from idpyoidc.server.configure import ASConfiguration
 from oauth2_client_conf import CLIENT_CONFIG
 from oauth2_client_conf import CLIENT_ID
 from oauth2_server_conf import SERVER_CONF
-
-KEYDEFS = [
-    {"type": "RSA", "key": "", "use": ["sig"]},
-    {"type": "EC", "crv": "P-256", "use": ["sig"]},
-]
-
-BASEDIR = os.path.abspath(os.path.dirname(__file__))
-
-
-def full_path(local_file):
-    return os.path.join(BASEDIR, local_file)
-
 
 # ================ Server side ===================================
 
@@ -73,7 +61,7 @@ client_conf['add_ons'] = {
         "function": "idpyoidc.client.oauth2.add_on.par.add_support",
         "kwargs": {
             'http_client': {
-                'class' : 'utils.EmulatePARCall'
+                'class': 'utils.EmulatePARCall'
             },
             'authn_method': 'client_secret_basic'
         }
