+++ demo/oauth2_add_on_pkce.py
@@ -1,6 +1,8 @@
 #!/usr/bin/env python3
 import os
 
+from common import BASEDIR
+from common import KEYDEFS
 from flow import Flow
 from idpyoidc.client.oauth2 import Client
 from idpyoidc.server import Server
@@ -8,13 +10,6 @@ from idpyoidc.server.configure import ASConfiguration
 from oauth2_client_conf import CLIENT_CONFIG
 from oauth2_client_conf import CLIENT_ID
 from oauth2_server_conf import SERVER_CONF
-
-KEYDEFS = [
-    {"type": "RSA", "key": "", "use": ["sig"]},
-    {"type": "EC", "crv": "P-256", "use": ["sig"]},
-]
-
-BASEDIR = os.path.abspath(os.path.dirname(__file__))
 
 
 def full_path(local_file):
