+++ demo/oauth2_cc.py
@@ -2,40 +2,17 @@
 """
 Displaying how Client Credentials works
 """
-import os
 
-from demo.flow import Flow
+from common import BASEDIR
+from common import KEYDEFS
+from common import SESSION_PARAMS
+from flow import Flow
 from idpyoidc.client.oauth2 import Client
 from idpyoidc.server import Server
 from idpyoidc.server.authz import AuthzHandling
 from idpyoidc.server.client_authn import verify_client
 from idpyoidc.server.configure import ASConfiguration
 from idpyoidc.server.oauth2.token import Token
-
-BASEDIR = os.path.abspath(os.path.dirname(__file__))
-
-
-def full_path(local_file):
-    return os.path.join(BASEDIR, local_file)
-
-
-KEYDEFS = [
-    {"type": "RSA", "key": "", "use": ["sig"]},
-    {"type": "EC", "crv": "P-256", "use": ["sig"]},
-]
-CRYPT_CONFIG = {
-    "kwargs": {
-        "keys": {
-            "key_defs": [
-                {"type": "OCT", "use": ["enc"], "kid": "password"},
-                {"type": "OCT", "use": ["enc"], "kid": "salt"},
-            ]
-        },
-        "iterations": 1,
-    }
-}
-
-SESSION_PARAMS = {"encrypter": CRYPT_CONFIG}
 
 SERVER_CONFIG = {
     "issuer": "https://example.net/",
