+++ demo/oidc_code_dyn_reg.py
@@ -1,20 +1,14 @@
 #!/usr/bin/env python3
 import os
 
-from demo.oidc_client_conf import CLIENT_CONFIG
-from demo.oidc_client_conf import CLIENT_ID
-from demo.oidc_server_conf import SERVER_CONF
+from common import BASEDIR
+from common import KEYDEFS
 from flow import Flow
 from idpyoidc.client.oidc import RP
 from idpyoidc.server import OPConfiguration
 from idpyoidc.server import Server
-
-KEYDEFS = [
-    {"type": "RSA", "key": "", "use": ["sig"]},
-    {"type": "EC", "crv": "P-256", "use": ["sig"]},
-]
-
-BASEDIR = os.path.abspath(os.path.dirname(__file__))
+from oidc_client_conf import CLIENT_CONFIG
+from oidc_server_conf import SERVER_CONF
 
 
 def full_path(local_file):
